refactor: Update package.json and useGameEngine.jsx, ViewerScreen.jsx, StreamScreen.jsx, and HostScreen.jsx components
Update the package.json file to use the latest version of the playroomkit dependency. In the useGameEngine.jsx file, add a new state variable wordList and a setter function setWordList. In the ViewerScreen.jsx component, add an input field and a button to set the word and update the wordList. In the StreamScreen.jsx component, display the wordList. In the HostScreen.jsx component, add a button to set up the game. These changes improve the functionality and user experience of the application.